requirements.txt
SQLAlchemy=1.4.45
gevent=22.10.2
greenlet=2.0.1 
zope.event=4.5.0 
zope.interface=5.5.2
psycopg2-binary=2.9.5
mysql-connector-python=8.0.31
protobuf=3.20.1
dacite=1.6.0
dict-to-dataclass=0.0.8
python-dateutil=2.8.1 
six=1.16.0
